+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Changed
+* `libuuid` dependency has been replaced with `modern-uuid`
+
 ## [1.16] - 2025-01-12
 
 ### Added